Clean up prototype with utils
