+++ app/controllers/document_controller.rb
@@ -30,11 +30,12 @@ class DocumentController < ApplicationController
     # The line below enables document caching for a month.
     expires_in 30.days, public: true
     send_file(
-      document.serve,
 
-############################ OCR ############################
+      document.serve,
+######################### non- OCR ###########################
       # "/Users/laurenjin/dev/appeals/caseflow/tmp/pdfs/non-PII.pdf",
       # x_sendfile: true,
+
 ############################ OCR ############################
       # "/Users/laurenjin/dev/appeals/caseflow/tmp/pdfs/OCR-non-PII.pdf",
       # x_sendfile: true,
