--- client/app/readerprototype/PdfPage.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import _ from 'lodash';
-
-const PdfPage = ({ isSelected }) => {
-
-}
-
-PdfPage.propTypes = {
-  isSelected: PropTypes.bool,
-};
-
-export default PdfPage;
