+++ client/app/readerprototype/components/ReaderToolbar.jsx
@@ -1,51 +1,18 @@
-import React from 'react';
 import PropTypes from 'prop-types';
-import { css } from 'glamor';
+import React from 'react';
 
-import Button from '../components/Button';
-import Link from '../components/Link';
+import Button from '../../components/Button';
+import Link from '../../components/Link';
+import { DownloadIcon } from '../../components/icons/DownloadIcon';
+import { ExternalLinkIcon } from '../../components/icons/ExternalLinkIcon';
+import { FitToScreenIcon } from '../../components/icons/FitToScreenIcon';
+import { LeftChevronIcon } from '../../components/icons/LeftChevronIcon';
+import { RotateIcon } from '../../components/icons/RotateIcon';
+import { SearchIcon } from '../../components/icons/SearchIcon';
+import DocumentCategoryIcons from '../../reader/DocumentCategoryIcons';
 
-import { DownloadIcon } from '../components/icons/DownloadIcon';
-import { LeftChevronIcon } from '../components/icons/LeftChevronIcon';
-import { ExternalLinkIcon } from '../components/icons/ExternalLinkIcon';
-import { FitToScreenIcon } from '../components/icons/FitToScreenIcon';
-import { RotateIcon } from '../components/icons/RotateIcon';
-import { SearchIcon } from '../components/icons/SearchIcon';
-import DocumentCategoryIcons from '../reader/DocumentCategoryIcons';
-
-const pdfWrapperSmall = 1165;
-
-const pdfToolbarStyles = {
-  openSidebarMenu: css({ marginRight: '2%' }),
-  toolbar: css({ width: '33%' }),
-  toolbarLeft: css({
-    '&&': { [`@media(max-width:${pdfWrapperSmall}px)`]: {
-      width: '18%' }
-    }
-  }),
-  toolbarCenter: css({
-    '&&': { [`@media(max-width:${pdfWrapperSmall}px)`]: {
-      width: '24%' }
-    }
-  }),
-  toolbarRight: css({
-    textAlign: 'right',
-    '&&': { [`@media(max-width:${pdfWrapperSmall}px)`]: {
-      width: '44%',
-      '& .cf-pdf-button-text': { display: 'none' } }
-    }
-  }),
-  footer: css({
-    position: 'absolute',
-    bottom: 0,
-    display: 'flex',
-    alignItems: 'center',
-    '&&': { [`@media(max-width:${pdfWrapperSmall}px)`]: {
-      '& .left-button-label': { display: 'none' },
-      '& .right-button-label': { display: 'none' }
-    } }
-  })
-};
+import { handleClickDocumentTypeLink, openDownloadLink } from '../documentUtil';
+import { pdfToolbarStyles } from '../layoutUtil';
 
 const ReaderToolbar = ({
   resetZoomLevel,
@@ -56,7 +23,9 @@ const ReaderToolbar = ({
   disableZoomOut,
   setZoomInLevel,
   disableZoomIn,
-  zoomLevel
+  zoomLevel,
+  // hideSidebar,
+  // toggleSidebar
 }) => {
   return <>
     <span {...pdfToolbarStyles.toolbar} {...pdfToolbarStyles.toolbarLeft}>
@@ -70,15 +39,10 @@ const ReaderToolbar = ({
         </Link>
       )}
     </span>
-
-    <span style={{ color: '#cc0000', fontWeight: 600 }}> PROTOTYPE!!!!! </span>
-
     <span {...pdfToolbarStyles.toolbar} {...pdfToolbarStyles.toolbarCenter}>
-      <span className="zoomContainer">
-
-      </span>
       <span className="category-icons-and-doc-type">
         <span className="cf-pdf-doc-category-icons">
+          <DocumentCategoryIcons doc={doc} />
         </span>
         <span className="cf-pdf-doc-type-button-container">
           <Link
@@ -86,12 +50,13 @@ const ReaderToolbar = ({
             ariaLabel="open document in new tab"
             target="_blank"
             button="matte"
+            onClick={() => handleClickDocumentTypeLink()}
             href={window.location.pathname.includes('prototype') ?
             `/reader/appeal${documentPathBase}/${doc.id}/prototype` :
             `/reader/appeal${documentPathBase}/${doc.id}`}
           >
             <h1 className="cf-pdf-vertically-center cf-non-stylized-header">
-              <span title="Open in new tab">{doc.docType}</span>
+              <span title="Open in new tab">{doc.type}</span>
               <span className="cf-pdf-external-link-icon"><ExternalLinkIcon /></span>
             </h1>
           </Link>
@@ -124,9 +89,43 @@ const ReaderToolbar = ({
         ariaLabel="fit to screen">
         <FitToScreenIcon />
       </Button>
-      {/* <Button name="rotation" /> */}
-      {/* <Button name="download" /> */}
-      {/* <Button name="search" /> */}
+      <Button
+        name="rotation"
+        classNames={['cf-pdf-button cf-pdf-spaced-buttons']}
+        // onClick={rotateDocument}
+        ariaLabel="rotate document"
+        disabled
+      >
+        <RotateIcon />
+      </Button>
+      <span className="cf-pdf-spaced-buttons">|</span>
+      <Button
+        name="download"
+        classNames={['cf-pdf-button cf-pdf-download-icon']}
+        onClick={() => openDownloadLink(doc)}
+        ariaLabel="download pdf">
+        <DownloadIcon />
+      </Button>
+      <Button
+        name="search"
+        classNames={['cf-pdf-button cf-pdf-search usa-search usa-search-small']}
+        ariaLabel="search text"
+        type="submit"
+        disabled
+        // onClick={toggleSearchBar()}
+      >
+        <SearchIcon />
+      </Button>
+      {/* {hideSidebar &&
+        (<span {...pdfToolbarStyles.openSidebarMenu}>
+          <Button
+            name="open sidebar menu"
+            classNames={['cf-pdf-button']}
+            onClick={() => toggleSidebar()}>
+            <strong> Open menu </strong>
+          </Button>
+        </span>)
+      } */}
     </span>
   </>;
 };
@@ -140,7 +139,9 @@ ReaderToolbar.propTypes = {
   disableZoomOut: PropTypes.bool,
   setZoomInLevel: PropTypes.func,
   disableZoomIn: PropTypes.bool,
-  zoomLevel: PropTypes.number
+  zoomLevel: PropTypes.number,
+  hideSidebar: PropTypes.bool,
+  toggleSidebar: PropTypes.func,
 };
 
 export default ReaderToolbar;
