+++ client/app/readerprototype/documentUtil.js
@@ -1,0 +1,57 @@
+import _ from 'lodash';
+import { CATEGORIES } from './readerConstants';
+
+export const isValidWholeNumber = (number) => {
+  return !isNaN(number) && number % 1 === 0;
+};
+
+export const validatePageNumber = (pageNumber, totalPages) => {
+  const RADIX = 10;
+  let pageNum = parseInt(pageNumber, RADIX);
+
+  if (!pageNum || !isValidWholeNumber(pageNum) ||
+    (pageNum < 1 || pageNum > totalPages)) {
+    return false;
+  }
+
+  return true;
+};
+
+export const selectedDocIndex = (props) => {
+  const selectedDocId = Number(props.match.params.docId);
+
+  return _.findIndex(props.allDocuments, { id: selectedDocId });
+};
+
+export const selectedDoc = (props) => (
+  props.allDocuments[selectedDocIndex(props)]
+);
+
+const getPrevDoc = (props) => _.get(props.allDocuments, [selectedDocIndex(props) - 1]);
+const getNextDoc = (props) => _.get(props.allDocuments, [selectedDocIndex(props) + 1]);
+
+export const getPrevDocId = (props) => _.get(getPrevDoc(props), 'id');
+export const getNextDocId = (props) => _.get(getNextDoc(props), 'id');
+
+// refactor - remove & use Redux
+export const getPageCount = () => {
+  // while !document.getElementById('pdfContainer') show loading
+  if (document.getElementById('pdfContainer')) {
+    return document.getElementById('pdfContainer').childElementCount;
+  }
+};
+
+export const handleKeyPress = (event) => {
+  if (event.key === 'Enter') {
+    document.getElementById(`canvasContainer-${event.target.value}`).scrollIntoView({ behavior: 'smooth' });
+  }
+};
+
+export const openDownloadLink = (doc) => {
+  window.analyticsEvent(CATEGORIES.VIEW_DOCUMENT_PAGE, 'download');
+  window.open(`${doc.content_url}?type=${doc.type}&download=true`);
+};
+
+export const handleClickDocumentTypeLink = () => {
+  window.analyticsEvent(CATEGORIES.VIEW_DOCUMENT_PAGE, 'document-type-link');
+}
