+++ client/app/readerprototype/readerConstants.js
@@ -1,0 +1,5 @@
+export const CATEGORIES = {
+  VIEW_DOCUMENT_PAGE: 'Document Viewer',
+  CLAIMS_FOLDER_PAGE: 'Claims Folder',
+  CASE_SELECTION_PAGE: 'Case Selection'
+};
